requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djoser==2.1.0
drf_yasg==1.20.0
psycopg2==2.9.1
python-decouple==3.4
xhtml2pdf==0.2.9
# python-dotenv==0.19.0
# asgiref==3.5.2
# autopep8==2.0.2
# distlib==0.3.6
# Django==3.2.18
# django-anymail==9.1
# django-crispy-forms==1.14.0
# django-js-asset==2.0.0
# django-rest-swagger==2.2.0
# django-templated-mail==1.1.1
# djangorestframework==3.14.0
# djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==4.8.0
# djoser==2.1.0
# drf-yasg==1.21.5
# gunicorn==20.1.0
# Jinja2==3.1.2
# lxml==4.9.2
# MarkupSafe==2.1.2
# oscrypto==1.3.0
# packaging==23.0
# Pillow==9.3.0
# pyjwt==2.6.0
# psycopg2==2.9.5
# pycodestyle==2.10.0
# pycparser==2.21
# pyHanko==0.17.2
# pyhanko-certvalidator==0.20.1
# pypdf==3.7.0
# pypng==0.20220715.0
# python-bidi==0.4.2
# python3-openid==3.2.0
# pytz==2022.7
# pytz-deprecation-shim==0.1.0.post0
# PyYAML==6.0
# qrcode==7.4.2
# reportlab==3.6.12
# requests==2.28.2
# requests-oauthlib==1.3.1
# ruamel.yaml==0.17.21
# ruamel.yaml.clib==0.2.7
# simplejson==3.18.3
# six==1.16.0
# sqlparse==0.4.3
# tomli==2.0.1
# typing_extensions==4.5.0
# tzdata==2022.6
# tzlocal==4.3
# uritemplate==4.1.1
# uritools==4.0.1
# urllib3==1.26.14
# virtualenv==20.16.6
# webencodings==0.5.1
# whitenoise==6.2.0
# xhtml2pdf==0.2.9
